requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
python-dotenv
passlib[bcrypt]  # Для хэширования паролей, если понадобится
